refactor(status): split aggregate.ts along service / subcheck boundary
The 309-line aggregate.ts becomes four focused modules:
- aggregateCommon.ts — bucketStart, percentile, mode, worstStatus, makeBucket,
  BucketAccumulator (shared by the two grouping passes)
- aggregateServices.ts — groupProbes (main-service bucketing)
- aggregateSubChecks.ts — groupSubChecks (per-sub-check bucketing)
- aggregate.ts — orchestrator: rebuildFine, mergeAggregate, runAggregate

Re-exports bucketStart from aggregate.ts so existing test imports keep
working. New aggregateCommon.test.ts unit-tests the pure helpers directly.
